Change WP CLI batch size flag to match AS core
Action Scheduler 1.6 is introducing a new WP CLI command which includes
a --batch-size flag. That serves the same purpose as the 'migrate'
command being provided here, but it was using a different name: --batch.

This updates it to make them consistent.

For more details on the AS core WP CLI command, see Prospress/action-scheduler#89